Home assistant seperation
