+++ include/websocket_handler.h
@@ -5,6 +5,12 @@
 #include <WebSocketsClient.h>
 #include <ArduinoJson.h>
 #include "config.h"
+#include "led_control.h"
+#include "entity_state.h"
+#include "utils.h"
+#include "animations.h"
+#include "secrets.h"
+#include "homeassistant_handler.h"
 
 extern WebSocketsClient webSocket;
 
@@ -17,11 +23,7 @@ extern QueuedMessage queuedMessages[MAX_QUEUED_MESSAGES];
 
 void initializeWebSocket();
 void reconnectWebSocket();
-void updateTimeAndCheckNightMode(const char* time_str);
-void toggleEntity(int x, int y);
 void webSocketEvent(WStype_t type, uint8_t * payload, size_t length);
-void subscribeToEntities();
 void queueWebSocketMessage(uint8_t* payload, size_t length);
 void processQueuedMessages();
-void sendBrightnessOrVolumeUpdate(const char* entity_id, int value, bool is_media_player);
 #endif // WEBSOCKET_HANDLER_H
